+++ bin/remove_hyperion.sh
@@ -6,22 +6,22 @@ PATH="/sbin:$PATH"
 
 #Check if HyperCon is logged in as root
 if [ $(id -u) != 0 ] && [ "$1" = "HyperConRemove" ]; then
-	echo '---> Critical Error: Please connect as user "root" through HyperCon' 
+	echo '---> Critical Error: Please connect as user "root" through HyperCon'
 	echo '---> We need admin privileges to remove your Hyperion! -> abort'
 	exit 1
 fi
 
 #Check, if script is running as root
 if [ $(id -u) != 0 ]; then
-	echo '---> Critical Error: Please run the script as root (sudo sh ./remove_hyperion.sh)' 
+	echo '---> Critical Error: Please run the script as root (sudo sh ./remove_hyperion.sh)'
 	exit 1
 fi
 
 #Welcome message
-echo '*******************************************************************************' 
-echo 'This script will remove Hyperion and its services' 
+echo '*******************************************************************************'
+echo 'This script will remove Hyperion and its services'
 echo '-----> Please BACKUP your hyperion.config.json if necessary <-----'
-echo 'Created by brindosch - hyperion-project.org - the official Hyperion source.' 
+echo 'Created by brindosch - hyperion-project.org - the official Hyperion source.'
 echo '*******************************************************************************'
 
 #Skip the prompt if HyperCon Remove
@@ -57,7 +57,7 @@ SERVICEC=1
 # Stop hyperion daemon if it is running
 echo '---> Stop Hyperion, if necessary'
 if [ $OS_OPENELEC -eq 1 ]; then
-    killall hyperiond 2>/dev/null
+    killall ambilightwifid 2>/dev/null
 elif [ $USE_INITCTL -eq 1 ]; then
 	/sbin/initctl stop hyperion 2>/dev/null
 elif [ $USE_SERVICE -eq 1 ]; then
@@ -77,8 +77,8 @@ if [ $USE_INITCTL -eq 1 ]; then
 elif [ $OS_OPENELEC -eq 1 ]; then
 	# Remove Hyperion from OpenELEC autostart.sh
 	echo "---> Remove Hyperion from OpenELEC autostart.sh"
-	sed -i "/hyperiond/d" /storage/.config/autostart.sh 2>/dev/null
-	sed -i "/hyperion-x11/d" /storage/.config/autostart.sh 2>/dev/null
+	sed -i "/ambilightwifid/d" /storage/.config/autostart.sh 2>/dev/null
+	sed -i "/ambilightwifi-x11/d" /storage/.config/autostart.sh 2>/dev/null
 elif [ $USE_SYSTEMD -eq 1 ]; then
 	# Delete and disable Hyperion systemd script
 	echo '---> Delete and disable Hyperion systemd script'
@@ -97,23 +97,22 @@ if [ $OS_OPENELEC -eq 1 ]; then
 	echo '---> Remove the OpenELEC Hyperion binaries and hyperion.config.json'
 	rm -rv /storage/hyperion 2>/dev/null
 	rm -v /storage/.config/hyperion.config.json 2>/dev/null
-else 	
+else
 	#Remove binaries on all distributions/systems (not OpenELEC)
-	echo "---> Remove links to the binaries"	
-	rm -v /usr/bin/hyperiond 2>/dev/null
-	rm -v /usr/bin/hyperion-remote 2>/dev/null
-	rm -v /usr/bin/hyperion-v4l2 2>/dev/null
-	rm -v /usr/bin/hyperion-dispmanx 2>/dev/null
-	rm -v /usr/bin/hyperion-x11 2>/dev/null
-	rm -v /usr/bin/hyperion-aml 2>/dev/null
+	echo "---> Remove links to the binaries"
+	rm -v /usr/bin/ambilightwifid 2>/dev/null
+	rm -v /usr/bin/ambilightwifi-remote 2>/dev/null
+	rm -v /usr/bin/ambilightwifi-v4l2 2>/dev/null
+	rm -v /usr/bin/ambilightwifi-dispmanx 2>/dev/null
+	rm -v /usr/bin/ambilightwifi-x11 2>/dev/null
+	rm -v /usr/bin/ambilightwifi-aml 2>/dev/null
 	rm -v /etc/hyperion.config.json 2>/dev/null
 	echo "---> Remove binaries"
 	rm -rv /opt/hyperion 2>/dev/null
 	rm -rv /etc/hyperion 2>/dev/null
 	rm -rv /usr/share/hyperion 2>/dev/null
 fi
-echo '*******************************************************************************' 
+echo '*******************************************************************************'
 echo 'Hyperion successful removed!'
-echo '*******************************************************************************'  
+echo '*******************************************************************************'
 exit 0
-